CORDA-3850 Add a per flow lock
Add a lock to `StateMachineState` that is used when interacting with a
flow. This prevents multiple threads from interacting with a flow at the
same time, therefore stopping race conditions and weird timing issues.

To take the lock call `FlowStateMachineImpl.withFlowLock`.

The lock should be taken at all points where interaction with a flow
starts. Currently this includes starting a transition
(`FlowStateMachineImpl.processEvent`) and kill flow (`SMM.killFlow`).