CORDA-3850 Remove unreachable code path in kill flow
A flow's `transientState` is nullable due to it being transient and not
being serialized when saving a checkpoint. By the time a flow is added
to the flows map, the `transientState` will *always* be set. Therefore,
we can assume that a flow has a `transientState` when it is in the map.

Therefore, in `killFlow` we do not need to check whether the state is
null as we can assume it has a value. This tidies up the code.