CORDA-3850 Do not re-set killed flag in processEvent
With the addition of the lock, it is no longer possible to have timing
issues when killing a flow. Therefore the state output as a result of a
transition can no longer interfere with state modified on another
thread.

The code in `processEvent` that maintained the killed flag in case the
flow was killed during a transition has been removed as it is not
needed.