CORDA-3777 Reload after every checkpoint
Enable reloading of a flow after every checkpoint is saved. This
includes reloading the checkpoint from the database and recreating the
fiber.

`reloadCheckpointAfterSuspend` is a field added to `Event.Suspend`. When
creating the event it checks the configuration to determine if it should
reload after processing the suspend event.

If it should reload, then it processes an `Event.RetryFlowFromSafePoint`
straight after completing the suspend. This handles the reloading.

This works because the flag on the suspend event, sets another flag on
the `StateMachineState`. When recreating the transient state
(`StateMachineState`), the same flag is set to false. This means when
the fiber is restarted (it resumes at the point after completing the
suspend event in `FlowStateMachineImpl.suspend`), it does not retry
again.

Without this switching between the flags the flow would be stuck in an
infinite loop of retrying.

A new staff member has been added to the flow hospital to force it to
ignore the `ReloadFlowFromCheckpointException` thrown if the flow fails
deserialization when reloading.